Change type for framerate

+++ src/common.cpp
@@ -740,9 +740,9 @@ void common_flip(struct game_obj *game_object)
 	if (!ff8)
 	{
 		time_t gametime;
-		long framerate = 60;
+		double framerate = 60.0;
 
-		if (mode->driver_mode == MODE_HIGHWAY) framerate = 30;
+		if (mode->driver_mode == MODE_HIGHWAY) framerate = 30.0;
 
 		do qpc_get_time(&gametime);
 		while (gametime > last_gametime && gametime - last_gametime < VREF(game_object, countspersecond * (1.0 / framerate)));
